Update patch set 1
Abandoned

Sorry this whole change is bogus. I thought that the person filing the bug had an actual case that failed. I went looking for something to be wrong and thought this might be wrong, but if this didn't work, then we would have lots of other problems.

Patch-set: 1
Status: abandoned