Update patch set 1
Patch Set 1:

> OK, then I don't think this public-libraries.txt file is being
 > generated on the bot. I suppose it would have to now, right?

Yes, there are 2 ways to do that - you can generate the reduced list of public-libraries.txt on buildbot, containing libc.so and libdl.so  and maybe libm? Or we can introduce a function to native loader resetting the list of public libraries and call it from dalvikvm.

I think the first way is cleaner.

Patch-set: 1