Update patch set 1
Patch Set 1: Code-Review+1

Elliott, Sami,

I'm facing an issue with my Android build including slot AB and a vendor partition.  The vendor partition cannot be mounted properly because the dm-android-verity driver receives an incorrect device path (vendor block device without the slot suffix).  The block device path is computed build time and included in the verity table and fs_mgr hand that value value directly to the kernel.

I don't think the kernel driver should do the substitution since it should not be aware of the slot AB logic.  This is not something that can be fixed build time neither.  IMHO, the appropriate place to handle it is fs_mgr (hence this patch).  Though, I'm really not sure this is the best.  What is your opinion ?

Thanks,

Jérémy

Patch-set: 1