Update patch set 1
Patch Set 1:

(2 comments)

Any chance we could ensure an ordering of event handling to preclude the race in the first place?  It would be nice if we could just ensure we have all the links known at creation time; that was part of the justification for allowing label-by-symlink in this situation.

Patch-set: 1
Label: Presubmit-Verified=0