Update patch set 1
Patch Set 1:

It would be nice to get the bounds and object-size sanitizers enabled, but object-size find a ridiculous number of benign bugs so it will be difficult.

The rest of -fsanitize=undefined is mostly not useful for hardening. For example, -fsanitize=null adds a branch to most dereferences to check for NULL, which is already going to fault due to mmap_min_addr. The only point of that pass is preventing the compiler from assuming that pointers are non-NULL due to usage and optimizing based on it.

It's probably a good idea to enable them on a case-by-case basis for production usage, while all of them can be used for debugging. Just to find a good performance trade-off and to deal with the benign issues that are uncovered bit-by-bit.

Patch-set: 1
Reviewer: Daniel Micay <1069238@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0