Update patch set 3
Patch Set 3: Code-Review+2

we already have a fairly significant diff between AOSP master and nyc-dev, don't we, so i guess it's fine to just do that separately in AOSP afterwards. as long as we have unique_fd in nyc-dev we'll be able to easily backport any future changes that require it.

(if this is one of the cherrypicks, though, add the Bug: lines now.)

Patch-set: 3
Reviewer: Elliott Hughes <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2