Update patch set 6
Patch Set 6:

(2 comments)

The old implementation of the android_atomic... operations had a full memory barrier in them, not just a release. I don't know when or if that changed (I don't see the assembly for the current implementation), but was not referring to the old version of this file.

Patch-set: 6