Update patch set 2
Patch Set 2:

I don't mind needing to parse data / binary data if it gives me access streaming notifications.  I currently have a bunch of Python code that is very heavy on "adb wait-for-device" to ensure that the device is still attached.  This becomes cumbersome when ADB is forwarded (e.g. over VPN / SSH) since the round-trip-time is large.  Needing to parse a small subset of the features ADB supports is much easier than supporting the entire ADB protocol (or even just enough of it to invoke this feature).

Patch-set: 2