Update patch set 3
Patch Set 3:

The expected output of the command:

logcatd
145
(return code 0)

I have not found any condition where we have returned an error, nor has anyone reported an error return value until this bug. It may be considered 'safe' to submit this prior to the fix in pgrep b/29092208 to unblock AOSP and master testing; letting the lack of logs collected by logpersist as the trigger, instead of a non-zero return code from logpersist.start, for bugs in this feature in the interim.

Y'all agree? (+2 if you do, and I will lift my -1 on Verified)

Patch-set: 3