Update patch set 1
Patch Set 1: Code-Review+1

(9 comments)

Ignore the NITs, they are notes to myself/mkayyash for a (set of) cleanup CLs to get this file to Android coding standard. I would prefer that the cleanup CL precede this one since this specific file is a fresh new file from Apr 7 this year landing in libcutils. Maybe fix strlcpy.c as well as yet another tab-filled straggler.

The only question relevant to this CL is if we want to also support leading '/' in canned files and source files alike and strip them out where necessary for strcmp's purpose?

Patch-set: 1
Label: Code-Review=+1