Update patch set 2
Patch Set 2:

> The issue is that in AOSP /init does the parsing of the vars, in N logcat does the parsing of the vars. N has _less_ expansions available as well.

But the only difference is in the `service ...` lines, which aren't changing here, right?  The handling of the setprop lines should be identical.

The change is low risk to cherry-pick back if we'd like to go down that regardless.

I see how/why those errors are being generated.  I don't think we ever had ever intended for it to be a valid use case to attempt to expand an empty property.  I'll have to think about how it could make sense to suppress that warning.

Patch-set: 2