Update patch set 2
Patch Set 2:

Hmm.  I understand what you're saying.  I'm not sure if adding a special case to setprop to indicate 'no op' is a good idea, as that would in theory prevent that same string from being used as a legitimate property value (and we don't have restrictions currently).

One other idea for removing these warnings, we could `exec` a script that handles the logic of "propagate only if not empty"

Patch-set: 2