Update patch set 2
Patch Set 2:

Current behavior is to not propagate on empty (error message notwithstanding), do not know if anyone else is taking advantage of that ... spawning to exec just seems to be adding some additional costs to the trampoline.

Another possibility, I like much better, could be to not allow on property:<property>=* to trigger on <empty> and require the caller to also specify property:<property>= to cover the <empty> case. This is cleaner because now you are _sure_ that ${<property>} is set in the =* case? Why else would someone use =* case if the possibility of the property could be set to an <empty> value?

Patch-set: 2