Update patch set 4
Patch Set 4: Verified+1

In unit testing, discovered:

EXT4-fs (sda31): mounted filesystem with ordered data mode. Opts: barrier=1,discard

did not correctly (?) identify 'EXT4-fs (sda31)' as a two-word tag.

Confirmed that this was the case before so this is a regression from an older adjustment. I see where the bug is, but that needs to be fixed in a different CL that is unrelated to this critical issue.

Patch-set: 4