Update patch set 1
Patch Set 1:

With this change on master, I've booted (on a 6P), and now I've just run the unit tests:

  adb shell /data/nativetest/libutils_tests/libutils_tests
    Running main() from gtest_main.cc
    [...]
    [==========] 120 tests from 12 test cases ran. (3530 ms total)
    [  PASSED  ] 120 tests.


In terms of a regression test, I'm sure one could be crafted.  However, this class is deprecated, according to the header file.  Also, these unit tests aren't automatically run (correct?).  So I would think effort would be better put into unit tests of the replacement of this class and/or getting these unit tests run in an automated manner.

But I won't +1 on Verified yet in case there's a strong opinion I should make a regression test for this before it's okay for submitting.

Patch-set: 1