Update patch set 1
Patch Set 1:

LGTM. Addressing a couple of Hans's comments 

(1) SharedBufferTest isn't really in an odd place. It's standard google3 / chromium / ART convention to place *unit* tests in the same directory as the implementation classes. Regression and integration tests are usually split out elsewhere.

(2) As for running it, you'll have to build and push it to your device by hand. These tests aren't run automatically or as a part of CTS for that matter but we should try and get there eventually. They may not be good CTS candidates because they don't test public API, but they can certainly be run on presubmit.

Patch-set: 1