Update patch set 1
Patch Set 1:

What's odd here is that the other libutils unit tests live in a tests subdirectory, and get built into a single executable, which seems far more convenient. I'd be inclined to also move SharedBufferTest to the tests directory. But the only thing that really seems broken is the inconsistency.

Patch-set: 1