Update patch set 1
Patch Set 1:

(2 comments)

One less syscall :-)

+2 for logd_writer.c, -1 for logd_test.cpp.

I would like to keep the blocking open for the gTest though; or check for an error return of EBUSY rather than skip to the next device node and mark _that_ as an error. It is a bit more complicated than that, an EBUSY return to the open means the node exists (which is also a test condition), but we can not discover if there is content so we must go on to the next node to see if there is content. Leaving it as-is is simpler than tracking existing and content separately. Pedantically nothing in this test confirms that the node is in fact the kernel logger device.

Patch-set: 1
Label: Presubmit-Verified=0