Update patch set 1
Patch Set 1: Autosubmit+1 Presubmit-Ready+1

We already removed a bunch of DHCPv4 code, yes, but I missed this.

It turns out we couldn't delete some of this code until legacy wifi was cleaned up, which has now happened.

After this change there will be one more to delete dhcpmsg et al., but that requires a bit more refactoring of some functions that are shared within it (but not in any header file, if I'm not mistaken).

Patch-set: 1
Label: Autosubmit=+1