Update patch set 1
Patch Set 1:

> As I understand it, as suggested by https://google.github.io/styleguide/cppguide.html#Static_and_Global_Variables
 > and http://go/totw/110 the best approach for global objects is to
 > instead use functions to access a singleton object, e.g.:
 > 
 > const string& GetKey() {
 > static string* key = new string("my secret key");
 > return *key;
 > }

I think the problem with this is that it results in false positives in valgrind. It seems like Real Google uses some other leak chcker that doesn't complain about this?

The base logging change does seem especially dangerous, though. Maybe this patch should just be abandoned and left as is?

Patch-set: 1