Update patch set 2
Patch Set 2:

(1 comment)

Issue that resurfaces with PS2: Any external NDK code that redefines the macros should do so ahead of android/log.h, or they can optionally #undef them if they want to override them after including this file. I believe this to be a small bump to the ecosystem, but want your feedback.

Patch-set: 2