Update patch set 2
Patch Set 2:

(3 comments)

> Actually, I have been thinking, there are use cases not supported
 > by this approach that mirror run-as. Sometimes, I have a vendor
 > service X that has privs Y, that needs to exec vendor binary W with
 > privs Q. I can set that all up in fsconfig and enable the
 > transitions and sand-boxing with SELinux.
 > 
 > The other thing that fsconfig lets me do, is then say, this binary
 > W should only be rx for the exact thing executing me. So if service
 > X was uid OEM_XXX, I could set up the mode on the binary W so only
 > OEM_XXX uids can execute it. It's a bit pedantic for sure
 > considering I could do something similar in SELinux, but I think
 > good DAC still has a place with SELinux on.

I'm not proposing getting rid of fs_config, if you read my commit message, I said "reduce the use of file capabilities". My objective with this patch is to get rid of the fact that AID_NET_ADMIN and AID_NET_RAW are both GIDs and grant capabilities. I want to drop those patches from Android kernels to reduce the delta between Android kernels and upstream.

As I explained on a previous comment, the existence of AID_NET_ADMIN and AID_NET_RAW is evidence that we already have two ways of setting capabilities. Moreover, the existence of AID_NET_RAW and AID_NET_ADMIN shows that it is useful to have capabilities that are inherited reasonable over execve(2). This patch aims primarily at getting rid of those UIDs, and also allow, if folks want to, to not have to *depend* on fs caps for things to work. As you point out, in certain cases fs caps might be more convenient, and fs_config might be necessary to set up DAC in a useful way.

Overall though the fact that when you proposed this in the first place the discussion didn't center around dropping the ancient AID-based capabilities doesn't mean we shouldn't do it now, does it?

Patch-set: 2