Update patch set 3
Patch Set 3:

(6 comments)

I actually tried not to combine them, naming the fields separately so that one could not pollute the other. At the beginning I thought they would be different, but when completed they were the same 'shape'. I still want the option to change the shape as I fear we are not done yet regarding the type. For instance, the cleanup is _not_ done on purpose, but we may wish to flag ones to clean up on service failure?

socket name is a <basename> while file path is the full and complete location. socketcon / filecon of course could just be context. If we make a base class we could do virtual for Create(), Publish() and Cleanup() in each and join them together at all phases in one list ...

I will look into how that will look ...

Patch-set: 3