Update patch set 3
Patch Set 3:

This is a leap forward as we drop root group. It is probably not hitting the pidns case (code optimization being what it is), so sorry for the misdirection. I will evaluate my options for how to remove root user at logd startup, but put that in a different set of CLs outside the scope of this patch stack. We are not adding more capabilities, they are already granted by the filesystem (I will double check that and update).

Patch-set: 3
Label: Verified=0