Update patch set 6
Patch Set 6: Code-Review+1

enh@ sez NO to inline ;->, so no other choice than:

#define _GNU_SOURCE at the top. Then after all includes add a:

#ifdef _LIBGEN_H
#warning "libgen.h must not be included"
#endif

(not sure if this piece of paranoia is really necessary, but it acts like a comment measuring the intent; a comment at basename would be the alternative).

Then switch to 'char * const argv[]' for main and parse as an additional measure to ensure no one touches the content. This will also cause basename to break the build if the wrong one is referenced.

then

me = basename(argv[0]);

and only then is it OK by me.

Patch-set: 6
Label: Code-Review=+1