Update patch set 1
Patch Set 1:

> No one uses it because it was inaccessible before, and everyone
 > implemented their own.
 > 
 > When we add the event log tag service (see bugz), we open the door
 > for at least exposing the ability to create binary event logs in
 > the (vapourware) Vendor NDK. More may come. We already have 3 spots
 > in the tree that use this C++ interface for creating binary event
 > logs; more may come.

that's my point. 3 users? that's basically 0.

3 copies is indeed the point at which we should stop repeating ourselves and factor stuff out, but libraries are free... make a new library for the folks that need this stuff.

this isn't something that's needed by everybody. big grab-bag libraries (cutils/utils) are a mistake.

(so is libbase, which is why we're trying to be relatively strict about it being "stuff that should have been in the STL, that almost everything containing more than one source file will need". which is why, no, something that will have a handful of users around the tree doesn't belong.)

Patch-set: 1