Update patch set 1
Patch Set 1: Code-Review+1

This change looks good to me; thanks for pushing it.

I had investigated this issue previously and was chiefly concerned about breaking the below case if the logic were to change:

(a = foo from boot)

on property:a=foo
  setprop b bar

on property:b=bar
  setprop c car

If property_triggers_enabled is set to 1 after the ("", "") triggers have run, when the second action never is called and 'c' is never set to 'car'.

This change solves the issue of the duplicate triggers without breaking the above case, as it only allows queue'ing property triggers immediately before it adds all actions from the ("", "") trigger to the queue.

Patch-set: 1
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1