Update patch set 6
Patch Set 6:

(2 comments)

Another factor of this code is to supply a base class for MapString that can eventually be extended to support std::string as a cache for runtime updates to the database. See https://docs.google.com/document/d/1cx21fBO4-aUGAjPDLTYnq-ukWoA1p3Ar35pMRhg33ew for some details.

I have the prototype for adding std::string to MapString in my private branch to keep any adjustments here honest. That is the reason we have the outstanding question to danalbert@ regarding the compatible hash function.

Patch-set: 6