Update patch set 6
Patch Set 6:

It should be noted that this provides a marginal value. It provides no testable or debuggable impact without the stack trace component. With the ability to grab a stack trace, then we are talking, as one can use it to pinpoint higher up the stack where the code has gone awry by the content in a bugreport.

Although it can help greatly reduce CPU cycles, or punish/slow, applications that are spamming the logs with identical content; who cares.

All useful information is present in logd as part of https://android-review.googlesource.com/#/c/314074, all reports that can aid pinpointing a bad application are all long-path via the logger daemon.

Patch-set: 6