Update patch set 3
Patch Set 3: Presubmit-Ready+1 Code-Review+1

Build errors were associated with an unrelated component, init, and do not appear to have anything to do with the change in the log header (selinux function that now takes two args not prototyped for the libselinux library?). Hitting Presubmit for a new determination.

Patch-set: 3
Label: Code-Review=+1
Label: Presubmit-Ready=+1