Update patch set 3
Patch Set 3: Code-Review-1

<editorial>
Not sure I am happy with the symbolic links as it is just creating more abstraction and wibbly wobbly timey wimey problems. The code that depends on it needs to be fixed _before_ you consider this CL. Our codebase is supposed to be a _good_ example of quality, not one hack piled on another.
</editorial>

Part of liblog missing from the move:

include/android/log.h  (Also NDK deliverable though)
include/private/android_logger.h

log/logd.h and log/logger.h are deprecated FYI, and point to log/log.h. I have mostly deprecated the users and am working through the fiddly bits. Given my editorial above, I may be chucking stones in a glass house, but I am doing so _because_ someone else did not perform the cleanup, instead taking yet another easy way out.

cutils:

cutils/log.h is deprecated and points to log/log.h which is part of liblog. (Debatably?) It should be part of liblog, not libcutils.

A fair amount of what lands in libcutils and libutils are header-only inlines. One should not have to reference the associated library in order to access them. utils/FastStrcmp.h for example would cause liblog to fail because it does _not_ ask for libutils to get it, had you not added the symbolic link. It should remain 'global' for this reason.

Patch-set: 3
Label: Code-Review=-1