Update patch set 2
Patch Set 2:

The issues surface every time selinux violations spam the logs, or when the logger is implicated in power consumption with selinux violations. We did another mitigation where we ignored the serial number and suppressed multiple identical avc message bodies that helped a little bit. We are always looking for other ways to keep the audit logging making sense, this _CL_ is IMHO not one of them though.

Patch-set: 2