Update patch set 2
Patch Set 2: Autosubmit+1 Code-Review+2

Ok, thanks, this change is pretty KISS anyways ...

As for AUDIT_SET, I _think_ it resets the sequence number when issued. If the logger crashes, for instance, the number resets. If this is the case when I prototype the code, how important is the preservation of the sequence number for analysis, or do we care? I could of course hold on to the sequence and fix it up in the logger, maybe even fix it up for the case of a crashing logger (code complexity notwithstanding)?

Patch-set: 2
Label: Autosubmit=+1
Label: Code-Review=+2
Label: Presubmit-Ready=0