Update patch set 2
Patch Set 2: Code-Review+2

Agreed for rationalization regarding init's single-threadedness and not waiting around for an ack. Agree on target uid/pid for this message as opposed to the normal source uid/pid in other avc reports. Looking at the INT-retry loop my lower-mind worries about the condition being indefinite in init, but realize interrupt is never a persistent condition of call failure so it is a nit with no basis in reality...

Patch-set: 2
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2