Update patch set 1
Patch Set 1:

(2 comments)

I'm surprised we don't have a pre-existing library which implements this kind of throttling. It's hard to judge by looking at the code if the algorithm you proposed is ideal. I wonder if having a ring-buffer for this might have less churn...  http://stackoverflow.com/questions/1407113/throttling-method-calls-to-m-requests-in-n-seconds

Patch-set: 1
Label: Verified=0