Update patch set 1
Patch Set 1:

(2 comments)

dequeue class backs queue by default and is as close to a ring buffer as one gets in stl. Implementing a static ring buffer would be open coding adding to the risk and reducing the maintainability, and may(?) cause an issue with trying to ascertain a high and low water mark as I have done here. In any case, I could implement a static produced-consumer ring buffer to back queue class; but I assure you given the comments regarding the class after the first block allocation for dequeue, they are likely one and the same.

Patch-set: 1