Update patch set 4
Patch Set 4:

hit post before I was done ...

libutils _and_ libsysutils are dangerous, maybe split them off so that C++ and C exports are handled separately from a review perspective.

Please also do not release utils/Log.h, questionable value in the hands of the vndk, and also one we would prefer never gets out.

We have been pushed to use <atomic> and <stdatomic.h>, utils/Atomic.h allows a dangerous bypass to the 64-bit alignment enforced in the global atomics for the 64-bit types and we need to deprecate this file in time.

Patch-set: 4