Update patch set 3
Patch Set 3:

This is ready for review. I've only included the parts that handles the Arm exidx stuff and everything needed to run that.

If you want to see how all of this goes together, look at this CL:

https://android-review.googlesource.com/#/c/245221/

This biggest chunks of the code are in ArmExidx and the tests for that code.

Patch-set: 3