Update patch set 1
Patch Set 1:

> > > Roland: I assume this requires to update the images on our
 > > testing
 > > > devices, right? Nothing else we could do to make it work with a
 > > > stock N?
 > >
 > > My assumption was that the strategy behind that CL would not
 > > require a change on the testing devices: we're adjusting to the
 > new
 > > debuggerd protocol (in master), which now requires a crash_dump
 > > binary.
 > 
 > Ah yes, building crash_dump on the buildbot might work. WDYT about
 > Josh's suggestion to just create a symlink (hence avoiding this
 > change with ANDROID_ROOT)? Would that work, or are there any other
 > hardcoded dependencies?

The symlink approach is a bit hack-ish IMO, as it is not really self-contained.  Also, it is not clear that this will solve all the issues: even if I managed to invoke `crash_dump` (with this CL), I was not able to get a stack trace in the logcat.

I'm trying the solution we discussed today, based on changing the signal handler in ART directly (my plan is to guard that with an explicit toggle), which would not require change outside ART. The longer (and better) plan is the chroot-based solution suggested by Josh; I'll create a bug for that.

Patch-set: 1
Reviewer: Roland Levillain <1052644@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0