Update patch set 3
Patch Set 3:

is this "ipa" service on all devices? or do you want to take advantage of the fact that you can now define a uid/gid in the service itself if nothing else needs to refer to it (and it doesn't actually need to be a constant across devices)?

Patch-set: 3