Update patch set 4
Patch Set 4:

(1 comment)

Holding off on the respin to add log/log_main.h into the vndk for this CL so that the top of stack Revert Revert can clear build and go in. At that point I will also update the Revert "Continue using the legacy NDK" to be relevant to this series.

Patch-set: 4