Update patch set 1
Patch Set 1: Code-Review-1

This is not really safe to do. It doesn't prevent the attack where partition is modified after this boot and the attacker then drops caches ("echo 3 > /proc/sys/vm/drop_caches")... and it's a really simple attack.

Also not sure that  "dm-verity on heavily accessed partitions leads to performance
slowdowns" is true ... IIRC the verity check happens before the data lands in the page cache. +drewry since he knows more (and can also comment on the performance cost in general). Is there a real-world use-case where the performance of dm-verity caused problems? If so maybe the answer is "get a better device", not "compromise on security".

Patch-set: 1
Label: Code-Review=-1