Update patch set 1
Patch Set 1: Code-Review-2

> > This is not really safe to do. It doesn't prevent the attack
 > where partition is modified after this boot
 > 
 > I don't believe this is a significant security concern. 

Not a significant concern, but it is not a no-op.  Additionally, I do not believe it makes sense to use dm-verity for this. It wastes disk space and at use, wastes ram and cpu.  It only makes sense to reuse dm-verity if we expect to share the partition and metadata with different products that do not need this flag.

Is that the case here? If not, the verification should just be a signature over the partition data.


> In case of
 > device compromise, dm-verity doesn't prevent malware from modifying
 > the partition. It does prevent the device from booting again next
 > time, which is also accomplished by verifying the entire partition
 > once at boot.

This is accurate but incomplete. dm-verity will not stop partition modification. However, it will stop that data from being used to abuse the _running_ system.  It will not relay an unverifiable block unless the kernel is compromised. That means that if a bug in a service that can write to the partition is exploited, it could be used for privilege escalation -- especially as kernel modules are enabled in future releases. It's easy enough to cause the caches to be dropped with memory use.

 > 
 > > Is there a real-world use-case where the performance of dm-verity
 > caused problems?
 > 
 > Yes, Android Wear. See b/32433608.

I'll look at this in a minute. The main cost after initial verification is if the page cache is dropped and the block needs to be rehashed. It will be one hash per block, so with arm instructions on modern chips (even low end) we should see good movement.

If not, i would encourage evaluating sha1. In this context, the risk of a break is quite low compared to the potential runtime abuse of verify-once.

I'd like to see more detail before we support this as part of the platform (and I'll look at the wear bug to see what was tried and not, etc!)

Thanks!
will

Patch-set: 1
Label: Code-Review=-2