Update patch set 1
Patch Set 1:

> I do not believe it makes sense to use dm-verity for this

Before AVB it made sense because we didn't support any other partition format, and we still wanted to use dm-verity in EIO mode instead of bricking the device completely if verification fails (there's no A/B in this case).

Otherwise, I agree that a simple signature would work the same way.

> However, it will stop that data from being used to abuse the _running_ system.

Modifying the partition already requires a kernel compromise because of SELinux, at which point dm-verity can easily be bypassed. If nothing else, malware can bind mount whatever it wants on top of the dm-verity partition.

> with arm instructions on modern chips (even low end) we should see good movement.

Turns out Wear devices are lower end than one might expect. Using a faster hash algorithm is an option, but such a change was too late for the Wear release schedule at that point.

Patch-set: 1