Update patch set 1
Patch Set 1:

I'm not sure I want to do this in general for all processes, because I'm worried about crashes in the linker not being reported but I believe this might be good enough for crash handling for seccomped processes. I'm operating under the assumption that it's unlikely for a seccomped process to be dynamically loading in more code after enabling a filter, and that it's good enough to have a dump in logcat but not an explicit tombstone (I can make it create a tombstone as well pretty easily, I just want to get this in first to make things at least slightly debuggable).

Do you think this is sufficient for media.codec/mediaserver, Marco?

Patch-set: 1