Update patch set 3
Patch Set 3:

> for e.g. this is the HAL IPACM will implement: https://android-review.googlesource.com/#/c/328689/
 > HAL definition is still a work in progress.

So HAL IPACM service will implement that interface, whose responsibility is it to provide that service? Is that service going to be a part of core Android? Is a 3rd party supposed to supply that service and it's supposed to run in a very particular uid?

Patch-set: 3
Reviewer: William C Roberts <1013433@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0