Update patch set 4
Patch Set 4: Presubmit-Ready+1

I have added some code for parsing 'writepid' which detects that writepid is used to write to a cpuset and if so, open the fd, etc as with "cpuset". Also issue a warning for such usage, so people know they need to switch to cpuset option.

Patch-set: 4
Label: Presubmit-Ready=+1