Update patch set 6
Patch Set 6:

(3 comments)

Yes, I am conflicted. cpuset sounds more to the point, easier to understand its place, but if it is to take over this duty, I want to see the cleanup of writepid usage in the tree _now_ as we have too many 'bad' examples.

Patch-set: 6
Label: Verified=0