Update patch set 1
Patch Set 1:

(1 comment)

Can this sysprop be retrieved in recovery mode as well? We hope to have that in recovery mode because the matching happens not only in OTA, but also when you download an OTA package manually and flash it in recovery mode.

Patch-set: 1
Reviewer: Yifan Hong <1121838@85c56323-6fa9-3386-8a01-6480fb634889>