Update patch set 13
Patch Set 13:

Adding my two cents here:

We're trying to formalize the way cpusets are handled in the system. Adding this command to the init language helps clarify the intent and furthers this formalism.

While I agree that the overlap with writepid could lead to confusion, I think the risk is overstated. It's okay to have some small overlap in functionality. Adding formal support for cpusets in the init language indicates that we are increasing the application of this feature, which is absolutely what we are doing -- as the number of cores increases so will the use of cpusets.

Patch-set: 13
CC: Corey Tabaka <1014369@85c56323-6fa9-3386-8a01-6480fb634889>