Update patch set 4
Patch Set 4:

> Cannot add PREUPLOAD.cfg because it's needs to under system/core,
 > same as .git folder. I might not have enough context to judge this,
 > as it impacts widely.

is there a way to write the PREUPLOAD.cfg so that it does nothing in directories that don't have a .clang-format?

(i do think we want this on for all of system/core/ at some point, but most directories don't have a .clang-format. though i think all of the heavily-modified ones do, minus logd, and from recent changes i assume salyzyn would like to move to clang-format too.)

Patch-set: 4