Update patch set 2
Patch Set 2:

Let me ask again, what is going on here? It's been dragging for three days already. The fix is simple, straightforward and within the constraints of the current logd code. If strnstr needs to be fixed, great, fix it with the appropriate testing, perf measurements, etc. but why the fix for the immediate problem at hand, the buffer overflow patch and other work depending on ti is stalled because of need of said broader investigation? How much longer it is going to take?

Patch-set: 2