Update patch set 3
Patch Set 3:

Practice sez otherwise, or we are dealing with a nomenclature issue (which I can accept).

# ( echo stdout ; echo stderr >&2 ) >/dev/null 2>&1
<crickets>
# ( echo stdout ; echo stderr >&2 ) 2>&1 >/dev/null
stderr
# ( echo stdout ; echo stderr >&2 ) >&2 2>/dev/null
stdout

The above output on bash, ash and ksh have redirects processed from right to left IMHO. As in the last case stderr is redirected to /dev/null. then stdout is _duped_ to a new stderr file descriptor (and thus is not in turn directed to /dev/null, since that has already been processed).

However:

# ( echo stdout ; echo stderr >&2 ) > out.txt | cat - >out2.txt
stderr
# ls -l out*
-rw-r-----  1 mark  admin  7 Feb 21 19:16 out.txt
-rw-r-----  1 mark  admin  0 Feb 21 19:16 out2.txt

We see the redirect to the left was processed before the pipe to the right. So processing was left to right.

# ( echo stdout ; echo stderr >&2 ) > out.txt 2>&1 | cat - >out2.txt
<crickets>
# ls -l out*
-rw-r-----  1 mark  admin  14 Feb 21 19:18 out.txt
-rw-r-----  1 mark  admin   0 Feb 21 19:18 out2.txt

So redirects processed from right to left, and pipe from left to right.

liblogcat redirection is processed for stdout first, then for stderr regardless of order.

Simplest phrase I can use to describe what is going on probably is: "liblogcat precedence order for redirects is stdout then stderr, which differs from shell processing" This should quell any debate?

Patch-set: 3
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>