Update patch set 3
Patch Set 3:

Maybe I should just state that 2>&1 must be last (even though it is not required) for shell compatibility (but 2>/dev/null can be anywhere).

[TL;DR]

Keep in mind _this_ patch is the one that gives us all the precedence of processing headaches. It was so simple when I just did stderr. It only gets complicated when we use the "2>&1", all other redirects are orthogonal.

I think I need to attack this differently. Stating that "2>&1" regardless of when/where it is placed will guarantee that stdout and stderr go to the same destination. State that this differs from the shell. We can skip precedence and ordering for any other redirect operations as they are normally orthogonal.

Ok, another test, I will force left to right for redirects using brackets:

# ( ( ( echo stdout ; echo stderr >&2 ) 2>&1 ) >/dev/null )
<crickets>
# ( echo stdout ; echo stderr >&2 ) 2>&1 >/dev/null
stderr

I do see, however, my comments are flawed in another regard: "We assume the _action_ of stderr 2>&2 redirection happens after stdout is dealt with" Scrub that too. That statement is not clear enough either. 2>&1 in the code sets the stdout_stderr flag, which means it joins its stderr output with stdout, wherever it is sent. Which means it is 'duped' to stdout first, then those both are sent to the same. The analog for bourne is actually ">FILENAME 2>&1", not the reverse as stated in the comment. But I can not skip this because this is the analog in bourne shell, and I want them to use this ordering when mixing them so that the 'script' can be handed to liblogcat or libc and do the same thing.

Patch-set: 3
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>