Update patch set 12
Patch Set 12:

(12 comments)

Sorry for not responding sooner, busy week!

A few notes, my big recommendation is to try to add some unittests for partition.cpp. I know there aren't a ton in fastboot yet, but new code should really have them, and this module is perfect for them - it's isolated from the rest of the code, and basically a black box that turns a string into a set of data structures.

Patch-set: 12
Reviewer: David Pursell <1078119@85c56323-6fa9-3386-8a01-6480fb634889>