Update patch set 1
Patch Set 1:

(3 comments)

The library currently is not thread or signal safe (not required by its current user, the logcat executable); but for this library to become generically more usable in the multitude of situations I have a long list of 'fixes'. Remember, this library started its life out _as_ the logcat executable and we are dozens of patches before we can bear fruit for the more esoteric users.

Patch-set: 1
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0