Update patch set 2
Patch Set 2: Code-Review-1

A typo had me place it in system/core, when I expected to land it in system/core/logd.

I only fixed the files that are being used by the dependent patches, so there is more disruption to come. I may mature this CL later to do this for all files in logd for reasons of purity and to rip the bandaid off in one blow (hence my -1)

Patch-set: 2
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1