Update patch set 18
Patch Set 18: Code-Review+2

Looks like everyone's comments are taken care of, so +2.

It'd be awesome if this code could be moved to ParseWritepid itself, though properties won't have been read from /system yet when it's running, so that's not possible.  In some other reality where the default wasn't a property, that may be an improvement.

Patch-set: 18
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2